requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
PyQt5>=5.15.0
pyqtgraph>=0.12.0
pyaudio>=0.2.11
pyyaml>=6.0

# Optional dependencies for BirdNET plugin
# Install one of the following:
birdnet>=0.1.7
# birdnetlib>=0.5.0
